+++ test/libsolidity/SMTCheckerTest.cpp
@@ -27,8 +27,8 @@ using namespace solidity::frontend;
 using namespace solidity::frontend::test;
 using namespace solidity::test;
 
-SMTCheckerTest::SMTCheckerTest(std::string const& _filename):
-	SyntaxTest(_filename, EVMVersion{}),
+SMTCheckerTest::SMTCheckerTest(std::string const& _filename, EVMVersion _evmVersion):
+	SyntaxTest(_filename, _evmVersion),
 	universalCallback(nullptr, smtCommand)
 {
 	auto contract = m_reader.stringSetting("SMTContract", "");