+++ test/libsolidity/syntaxTests/constants/initialization/math_builtin_precompile_based_const_args_future.sol
@@ -1,0 +1,21 @@
+uint8 constant k = 1;
+bytes constant data = hex"ffff";
+
+bytes32 constant sGlobal = sha256(hex"ffff");
+bytes32 constant sConstArgGlobal = sha256(data);
+address constant addrGlobal = ecrecover("1234", k, "0", abi.decode("", (bytes2)));
+bytes20 constant ripemdGlobal = ripemd160(hex"ffff");
+bytes20 constant ripemdConstArgGlobal = ripemd160(data);
+
+contract A {
+    bytes32 constant s = sha256(hex"ffff");
+    bytes32 constant sConstArg = sha256(data);
+    address constant addr = ecrecover("1234", k, "0", abi.decode("", (bytes2)));
+    bytes20 constant ripemd = ripemd160(hex"ffff");
+    bytes20 constant ripemdConstArg = ripemd160(data);
+}
+// ====
+// EVMVersion: =@future
+// ----
+// TypeError 8349: (181-232): Initial value for constant variable has to be compile-time constant.
+// TypeError 8349: (478-529): Initial value for constant variable has to be compile-time constant.