+++ test/libsolidity/syntaxTests/nameAndTypeResolution/496_builtin_ecrecover_reject_gas_future.sol
@@ -1,0 +1,9 @@
+contract C {
+    function f() public {
+        ecrecover.gas();
+    }
+}
+// ====
+// EVMVersion: =@future
+// ----
+// TypeError 9582: (47-60): Member "gas" not found or not visible after argument-dependent lookup in function (bytes32,uint8,bytes32,bytes32) view returns (address).